software-pc/requirements.txt
opencv-python
numpy
flask
Pillow
requests
